Move data resize logic to dataset file to get rid of numpy reference
